src/requirements.txt
pymongo
boto3
paho-mqtt
pyjwt
pymongo-inmemory
joblib
